SCRUM-31 Unificar archivos de configuracion
Se unificaron data-source y typeorm bajo el nombre typeorm
Se elimino el archivo data-source
